refactor(auth): hide Supabase coordinator helpers
